requirements.txt
Django==1.5.2
MySQL-python==1.2.4
django-tastypie==0.10.0
mimeparse==0.1.3
mongoengine==0.7.10
mysql-connector-python==1.0.9
pymongo==2.5
python-dateutil==2.1
six==1.3.0
wsgiref==0.1.2
